+++ bench_test.go
@@ -9,7 +9,7 @@ import (
 func BenchmarkPick(b *testing.B) {
 	// incoming request
 	var buf bytes.Buffer
-	buf.WriteString(`{"Name":"John Doe"}`)
+	buf.WriteString(`{"Name":"John Doe", Width: 231}`)
 
 	r := httptest.NewRequest("POST", "/person/123?group=aliens&copies=10&flag=true", &buf)
 	r.Header.Set("content-type", "application/json")